requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
pandas==2.1.3
google-generativeai==0.3.1
python-multipart==0.0.6
pydantic==2.5.2
numpy==1.26.2
scikit-learn==1.3.2
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1 
pydantic[email]
aiosqlite==0.19.0
alembic==1.12.1
pytest==7.4.3
httpx==0.25.2
aiosqlite
alembic